requirements.txt
selenium~=4.31.0
pytest~=8.3.5
locust~=2.34.0
requests~=2.32.3
PyJWT~=2.10.1
flask~=3.1.0
gunicorn